+++ Package.swift
@@ -1,25 +1,23 @@
-// swift-tools-version: 5.7
+// swift-tools-version: 5.8
+// The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
 
 let package = Package(
     name: "MyCLI",
-    products: [
-        .executable(name: "MyCLI", targets: ["MyCLI"])
-    ],
     dependencies: [
-        .package(url: "https://github.com/apple/example-package-figlet.git", branch: "main"),
-        .package(url: "https://github.com/apple/swift-argument-parser", from: "1.0.0")
+      .package(url: "https://github.com/apple/example-package-figlet", branch: "main"),
+      .package(url: "https://github.com/apple/swift-argument-parser", from: "1.0.0"),
     ],
     targets: [
+        // Targets are the basic building blocks of a package, defining a module or a test suite.
+        // Targets can depend on other targets in this package and products from dependencies.
         .executableTarget(
             name: "MyCLI",
             dependencies: [
                 .product(name: "Figlet", package: "example-package-figlet"),
                 .product(name: "ArgumentParser", package: "swift-argument-parser"),
-            ]),
-        .testTarget(
-            name: "MyCLITests",
-            dependencies: ["MyCLI"]),
+            ],
+            path: "Sources"),
     ]
 )
