--- Sources/MyCLI/MyCLI.swift
@@ -1,24 +1,0 @@
-//===----------------------------------------------------------------------===//
-//
-// This source file is part of the Swift open source project
-//
-// Copyright (c) 2023 Apple Inc. and the Swift project authors
-// Licensed under Apache License v2.0 with Runtime Library Exception
-//
-// See http://swift.org/LICENSE.txt for license information
-// See http://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
-//
-//===--------------------------------------------------------------------
-
-import ArgumentParser
-import Figlet
-
-@main
-struct FigletTool: ParsableCommand {
-  @Option(help: "Specify the input")
-  public var input: String
-
-  public func run() throws {
-    Figlet.say(self.input)
-  }
-}
